ISSUE #5889 - Fixed datepicker styles
