--- frontend/src/v5/ui/controls/inputs/datePicker/baseCalendarPicker/baseCalendarPicker.component.tsx
@@ -1,135 +1,0 @@
-/**
- *  Copyright (C) 2022 3D Repo Ltd
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-import { useEffect, useRef, useState } from 'react';
-import { FormInputProps } from '@controls/inputs/inputController.component';
-import { formatMessage } from '@/v5/services/intl';
-import { TextField } from './baseCalendarPicker.styles';
-import { formatDateTime } from '@/v5/helpers/intl.helper';
-import { ClearIcon, DateTimePicker } from '@mui/x-date-pickers';
-
-export type BaseCalendarPickerProps = FormInputProps & {
-	defaultValue?: Date;
-	PickerComponent: any;
-	placeholder?: string;
-};
-
-export const BaseCalendarPicker = ({
-	disabled,
-	defaultValue,
-	PickerComponent,
-	helperText,
-	error,
-	required,
-	value,
-	onChange,
-	onBlur,
-	placeholder,
-	...props
-}: BaseCalendarPickerProps) => {
-	const [open, setOpen] = useState(false);
-	const inputRef =  useRef(null);
-	const changeAborted =  useRef(false);
-
-	const closePicker = () => setOpen(false);
-
-	const preventPropagation = (e) => {
-		if (e.key !== 'Escape') {
-			changeAborted.current = true;
-			e.preventDefault();
-			e.stopPropagation();
-		}
-	};
-
-	const handleClick = (e) => {
-		if (disabled) return;
-		preventPropagation(e);
-		if (!open) {
-			setOpen(true);
-		}
-	};
-
-	useEffect(() => {
-		if (inputRef.current) {
-			inputRef.current.value = value ? formatDateTime(value) : null;
-		}
-	}, [inputRef.current, value]);
-
-	return (<DateTimePicker
-		{...props}
-		closeOnSelect
-		localeText={{ 
-			clearButtonLabel:  formatMessage({
-				id: 'calendarPicker.clearButtonLabel',
-				defaultMessage: 'Clear' }),
-		}} 
-		slots={{
-			textField: (textFieldProps) => {
-				return (<TextField
-					{...textFieldProps}
-					onKeyDown={(e) => e.preventDefault()}
-					placeholder={placeholder ?? formatMessage({
-						id: 'calendarPicker.placeholder',
-						defaultMessage: 'Choose a date',
-					})}
-
-					slotProps={{
-						input: textFieldProps.InputProps,
-					}}
-
-					value={undefined}
-					inputRef={inputRef}
-					onClick={handleClick}
-				/>);
-			},
-			clearIcon: ClearIcon, 
-		}}
-
-		slotProps={{ 
-			actionBar: { actions: ['clear', 'accept', 'cancel'] },
-			desktopPaper: {
-				onMouseLeave: () => {
-					changeAborted.current = true;
-				},
-				onMouseMove: () => {
-					changeAborted.current = false;
-				},
-			},
-		}}
-
-		enableAccessibleFieldDOMStructure={false} 
-		open={open}
-
-		onChange={(val) => { 
-			inputRef.current.value = formatDateTime(val);
-		}}
-
-		onAccept={(newValue) => {
-			if (!changeAborted.current) {
-				onChange?.(newValue);
-				onBlur?.();
-			}
-			closePicker();
-		}}
-
-		onClose={() => {
-			inputRef.current.value = formatDateTime(value);
-			closePicker();
-		}}
-
-	/>);
-
-};
